🔨 refactor : staffIntro 파일명 변경, post-test, post-update-test 파일의 이름에 test제거,  컴포넌트끼리의 이름 통일( Compo로 통일)
