+++ web/src/app/(Layout)/components/EditorCompo.tsx
