+++ web/src/app/(Layout)/components/StaffIntro.tsx
