+++ web/src/app/(Layout)/post/[category]/page.tsx
@@ -1,5 +1,5 @@
 import HtmlDocs from "../../components/HtmlDocs";
-import EditorComponent from "../../components/EditorComponent";
+import EditorComponent from "../../components/EditorCompo";
 
 interface WriteProps{
   params:{
